requirements.txt
asgiref==3.10.0
channels==4.3.1
channels_redis==4.3.0
colorama==0.4.6
Django==5.2.7
django-cors-headers==4.9.0
django-debug-toolbar==6.1.0
django-filter==25.2
daphne==4.0.0
iniconfig==2.3.0
msgpack==1.1.2
packaging==25.0
pillow==12.0.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.2
python-dotenv==1.2.1
redis==7.0.1
sqlparse==0.5.3
tzdata==2025.2
